Move `isWatchFileSystem` to StartParameterInternal
so it is not exposed via the public API.
