+++ subprojects/core-api/src/main/java/org/gradle/StartParameter.java
@@ -91,7 +91,6 @@ public class StartParameter implements LoggingConfiguration, ParallelismConfigur
     private boolean refreshDependencies;
     private boolean buildCacheEnabled;
     private boolean buildCacheDebugLogging;
-    private boolean watchFileSystem;
     private boolean configureOnDemand;
     private boolean continuous;
     private List<File> includedBuilds = new ArrayList<>();
@@ -716,26 +715,6 @@ public class StartParameter implements LoggingConfiguration, ParallelismConfigur
      */
     public void setBuildCacheDebugLogging(boolean buildCacheDebugLogging) {
         this.buildCacheDebugLogging = buildCacheDebugLogging;
-    }
-
-    /**
-     * Whether watching the file system for faster up-to-date checking is enabled.
-     *
-     * @since 6.5
-     */
-    @Incubating
-    public boolean isWatchFileSystem() {
-        return watchFileSystem;
-    }
-
-    /**
-     * Whether watching the file system for faster up-to-date checking is enabled.
-     *
-     * @since 6.5
-     */
-    @Incubating
-    public void setWatchFileSystem(boolean watchFileSystem) {
-        this.watchFileSystem = watchFileSystem;
     }
 
     /**
