+++ subprojects/core/src/main/java/org/gradle/execution/ProjectExecutionServices.java
@@ -20,6 +20,7 @@ import org.gradle.StartParameter;
 import org.gradle.api.execution.TaskActionListener;
 import org.gradle.api.execution.TaskExecutionListener;
 import org.gradle.api.execution.internal.TaskInputsListeners;
+import org.gradle.api.internal.StartParameterInternal;
 import org.gradle.api.internal.cache.StringInterner;
 import org.gradle.api.internal.changedetection.TaskExecutionModeResolver;
 import org.gradle.api.internal.changedetection.changes.DefaultTaskExecutionModeResolver;
@@ -136,15 +137,15 @@ public class ProjectExecutionServices extends DefaultServiceRegistry {
         TaskSnapshotter taskSnapshotter,
         WorkExecutor<ExecutionRequestContext, CachingResult> workExecutor
     ) {
-
-        ExecuteActionsTaskExecuter.VfsInvalidationStrategy vfsInvalidationStrategy = VirtualFileSystemServices.isPartialInvalidationEnabled(startParameter)
+        StartParameterInternal startParameterInternal = (StartParameterInternal) startParameter;
+        ExecuteActionsTaskExecuter.VfsInvalidationStrategy vfsInvalidationStrategy = VirtualFileSystemServices.isPartialInvalidationEnabled(startParameterInternal)
             ? ExecuteActionsTaskExecuter.VfsInvalidationStrategy.PARTIAL
             : ExecuteActionsTaskExecuter.VfsInvalidationStrategy.COMPLETE;
 
         // TODO: The incubation message should be printed in VirtualFileSystemServices.
         //   The problem is that `RootBuildLifecycleListener.afterStart` is called to early to have the system properties from gradle.properties available
         //   We log the message now here as a workaround.
-        if (vfsInvalidationStrategy == ExecuteActionsTaskExecuter.VfsInvalidationStrategy.PARTIAL && !startParameter.isWatchFileSystem()) {
+        if (vfsInvalidationStrategy == ExecuteActionsTaskExecuter.VfsInvalidationStrategy.PARTIAL && !startParameterInternal.isWatchFileSystem()) {
             IncubationLogger.incubatingFeatureUsed("Partial virtual file system invalidation");
         }
         TaskExecuter executer = new ExecuteActionsTaskExecuter(
